Add missing dependencies and resolve circular dependencies. Change: 140079490
