+++ tensorflow/python/platform/benchmark.py
@@ -29,8 +29,6 @@ import six
 
 from tensorflow.core.protobuf import config_pb2
 from tensorflow.core.util import test_log_pb2
-# timeline is outside of the platform target, but is brought in by the target
-# ":platform_test", which also brings in ":platform" (and thus this library).
 from tensorflow.python.client import timeline
 from tensorflow.python.platform import app
 from tensorflow.python.platform import gfile
